Project delete -> Confirmation implemented

+++ src/main/java/ch/ibw/knxgac/View/KnxGacApplication.java
@@ -316,34 +316,41 @@ public class KnxGacApplication extends Application {
             guiProject.btnDelete.setOnAction(new EventHandler<ActionEvent>() {
                 @Override
                 public void handle(ActionEvent actionEvent) {
-                    Project project = new Project(KnxGacApplication.currentProjectID);
-                    try {
-                        // delete the actual project
-                        controller.deleteObject(project);
-                        // update the projects from db
-                        guiProject.updateProjects();
-                        // empty the ComboBox
-                        guiProject.selectProject.getItems().clear();
-                        // add the project-items new to the ComboBox
-                        guiProject.selectProject.getItems().addAll(guiProject.projectItems());
-                        KnxGacApplication.currentProjectName = "";
-                        KnxGacApplication.currentProjectID = 0;
-                        // Windowtitle
-                        // actual Project in stage.title
-                        String title = stage.getTitle();
-                        int pos = title.indexOf("  /  ");
-                        if(pos>0)
-                            title = title.substring(0, pos);
-                        stage.setTitle(title + "  /  Kein Projekt gewählt");
-                        // disable tabs
-                        tabMaingroup.setDisable(true);
-                        tabMiddlegroup.setDisable(true);
-                        tabAddress.setDisable(true);
-                        // disable buttons
-                        guiProject.btnDelete.setDisable(true);
-                        guiProject.btnExport.setDisable(true);
-                    } catch (SQLException e) {
-                        e.printStackTrace();
+                    Alert d = new Dialog().getConfirmation(
+                            "Löschen bestätigen",
+                            "Projekt wirklich löschen?",
+                            "Wollen Sie das Projekt endgültig löschen?");
+                    d.showAndWait();
+                    if(d.getResult().getText().equalsIgnoreCase("ok")) {
+                        Project project = new Project(KnxGacApplication.currentProjectID);
+                        try {
+                            // delete the actual project
+                            controller.deleteObject(project);
+                            // update the projects from db
+                            guiProject.updateProjects();
+                            // empty the ComboBox
+                            guiProject.selectProject.getItems().clear();
+                            // add the project-items new to the ComboBox
+                            guiProject.selectProject.getItems().addAll(guiProject.projectItems());
+                            KnxGacApplication.currentProjectName = "";
+                            KnxGacApplication.currentProjectID = 0;
+                            // Windowtitle
+                            // actual Project in stage.title
+                            String title = stage.getTitle();
+                            int pos = title.indexOf("  /  ");
+                            if (pos > 0)
+                                title = title.substring(0, pos);
+                            stage.setTitle(title + "  /  Kein Projekt gewählt");
+                            // disable tabs
+                            tabMaingroup.setDisable(true);
+                            tabMiddlegroup.setDisable(true);
+                            tabAddress.setDisable(true);
+                            // disable buttons
+                            guiProject.btnDelete.setDisable(true);
+                            guiProject.btnExport.setDisable(true);
+                        } catch (SQLException e) {
+                            e.printStackTrace();
+                        }
                     }
                 }
             });
